service--text-extractor/pyproject.toml
[project]
name = "text-extractor"
version = "0.0.1"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.13"
dependencies = [
    "pydantic~=2.12.5",
    "pydantic-settings~=2.12.0",
    "loguru~=0.7.2",
    "dishka~=1.8.0",
    "mmar-mapi~=1.6.0",
    "mmar-ptag~=1.1.2",
    "mmar-utils~=1.1.18",

    "pytesseract==0.3.13",
    "pypdf==4.1.0",
]

[dependency-groups]
dev = [
    "fire~=0.7.0",
    "python-dotenv~=1.0.0",
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
pythonpath = [ "src" ]


[tool.ruff]
line-length = 120